Feature/modulize botfunctions (#1)
* moved test code to test dir

* move main app to own dir

* update pycache path

* rename dryrun to passive mode

* add new path to gitignore

* Restructured to individual module files

* create function to retrieve access_token

* Restructured functions to moduralized layout

* listen to function rand instead of random

* add __pycache__ paths to gitignore

* Update Dockerfile to reflect code structure

* Restructured code

* small code layout update

* change path Dockerfile

* move requirements.txt

* sourcing environment variables

* Adding req and opt  variables to Dockerfile

* relocate docker client initialized

* Adding Pycharm files

* Check gif function for valid response

* Handle gif return messages.

* remove build from repo

* adding pytest

* Moving modules to examples dir

* Moving modules to examples dir

* adding empty __init__.py for structure

* unittest for message class

* pytest for message class

* adding setup.py with project metadata

* moved test modules to example directory

* change build path for exclusion

* remove not required parentheses.

* update requirements with pytest modules

* adding pytest testcases

* update setup with new package requirements

* flake8 run

* chagne import statement to use in testcases

* update gitignroe with coveragedb and cache

* removed empty __init__.py

* change import to relative import

* Moved pytest case to do proper testing

* removed old tests dir

* removed comment

* rewrite of flip test

* update version and fix typo

* add oauth note to twitch service

* adding new enviroment vars to Dockerifle

* sourcing apikey from environment variable

* changed APIKEY variables

* sourcing APIKEY informatin for environment variable

* change twitch info to oauth2

* Update gitignore with test setenv script

* change twitch module to use credentials from env

* only query gnews api when GNEWS_APIKEY is set.

* flake8 fix.

* gif module sourcing environment variable APIKEY

* gnews module sourcing environment variable APIKEY

* twitch module credentials now depending on environment vars

* deduplicate retrieval of access_token

* fix utf-8 decoding replace unicode chars in parsing

* removed comment

* changed layout and added additional tests

* update regexp match for bored and chuck functions

* adding pytest-dotenv

* adding testemoji module

* adding testemoji module

* fixed the gif testcase and enabled test and testemoji

* adding pytest.ini

* renamed twitch test function

* Centralize author and version data in metadata module

* add coverage directory to gitignore

* update readme with development information
