+++ README.md
@@ -2,7 +2,7 @@
 
 
 ## Create intial config
-```
+```buildoutcfg
 mkdir $HOME/signal
 docker run -v $HOME/signal:/config --rm -it pblaas/signal-cli:latest link
 ```
@@ -12,18 +12,18 @@ Paste entire tsdevice:/? string in https://www.nayuki.io/page/qr-code-generator-
 
 ## Start messaging
 After device is connected you can start sending messages:
-```
+```buildoutcfg
 docker run -v $HOME/signal:/config --rm -it pblaas/signal-cli:latest -u YOURREGISTEREDNR send RECEIVER -m "your message"
 ```
 
 
 ## show CLI help
-```
+```buildoutcfg
 docker run -v $HOME/signal:/config --rm -it pblaas/signal-cli:latest -h
 ```
 
 ## having some fun
-```
+```buildoutcfg
 echo `curl --silent https://api.chucknorris.io/jokes/random | jq '. | .value'` | docker run -v $HOME/signal:/config --rm -i signal:dev -u +31630030905 send --g "2SElh8hai/NQTSNaBOpHKBc0BbYE90l1iQyXAQzfeoE="
 ```
 
@@ -42,7 +42,7 @@ echo `curl --silent https://api.chucknorris.io/jokes/random | jq '. | .value'` |
 * !names - random names based on python haikunator implementation [DONE]
 * !launch - return lanched message and send source a image of impact.
 * !trivia - return trivia questions [DONE] [API]
-* !twitch - Return twitch and game related info [DONE] [API] [APIKEY]
+* !twitch - Return twitch and game related info [DONE] [API] [OAUTH2]
 * !bored - Returns random activities [DONE]  [API]
 * !dog - Return short message with Emoji [DONE] 
 * !winamp - Returns winamp slogan with Emoji [DONE]
@@ -82,7 +82,7 @@ To set local executor mode change boolean on the top of the app.py to:
 `SIGNALEXECUTORLOCAL = True`
 
 To start the bot run:
-```
+```buildoutcfg
 docker run -v $HOME/signal:/config --rm -it pblaas/signalbot
 ```
 
@@ -94,7 +94,7 @@ To set non-local executor mode change boolean on the top of the app.py to:
 `SIGNALEXECUTORLOCAL = False`
 
 To start the bot run:
-```
+```buildoutcfg
 pip3 install -r requirements
 python3 app.py
 ```
@@ -106,7 +106,7 @@ The first step into using this app is making sure Signal CLI has a proper user p
 ### Linking
 
 To start a linking process with Signal CLI one should provide the link flag to signal-cli
-```
+```buildoutcfg
 signal-cli link
 ```
 
@@ -116,12 +116,39 @@ Use your Signal APP e.g on your phone to add additional devices by scanning the 
 ### Registering
 
 If you are not linking the bot to an existing account you can register the bot with the register flag.
-```
+```buildoutcfg
 signal-cli -u YOURPHONENUMBER register
 ```
 
 You will then receive a SMS message with validation code.
-```
+```buildoutcfg
 signal-cli -u YOURPHONENUMBER verify CODE
 ```
 
+# Development
+
+Additional bot functions should be added to the signalbot/functions directory. The modules should contain a class
+and one or multiple methods. The class then should also be added to the __init__.py file in the functions directory.
+
+Next the new class should be added to the SwitchCase class in botfunctions.py so its content can be inherited. 
+Reviewing some existing functions should give a good idea on how the bot can be extended.
+
+
+Testcases are written in Pytest and utilize a .env file with the required variables:
+```buildoutcfg
+READY=False
+DEBUG=True
+SIGNALEXECUTORLOCAL=False
+REGISTEREDNR="+316"
+GIPHY_APIKEY=""
+GNEWS_APIKEY=""
+TWITCH_CLIENTID=""
+```
+## Running the Pytest testsuite
+```buildoutcfg
+pytest signalbot/
+```
+## Run Pytest code coverage check
+```buildoutcfg
+pytest --cov-report html:cov_html --cov=signalbot signalbot/
+```
