+++ examples/emoij.py
