+++ examples/gifs.py
