+++ examples/gnews.py
