+++ examples/scatchpad.txt
