+++ examples/trivia.py
