+++ pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+env_files =
+    .env
+    .test.env
+    .deploy.env
