+++ signalbot/functions/__init__.py
@@ -1,0 +1,21 @@
+"""Functions package file."""
+
+from .help import Help
+from .test import Test
+from .testemoji import TestEmoji
+from .version import Version
+from .gif import Gif
+from .chuck import Chuck
+from .flip import Flip
+from .rand import Rand
+from .haiku import Haiku
+from .me import Me
+from .bored import Bored
+from .trivia import Trivia
+from .gnews import Gnews
+from .twitch import Twitch
+from .hn import Hn
+from .dog import Dog
+from .winamp import Winamp
+
+__all__ = ('Help', 'Test', 'TestEmoji', 'Version', 'Gif', 'Chuck', 'Flip', 'Rand', 'Haiku', 'Me', 'Bored', 'Trivia', 'Gnews', 'Twitch', 'Hn', 'Dog', 'Winamp')
