+++ signalbot/functions/help.py
@@ -1,0 +1,25 @@
+"""Help module."""
+
+
+class Help:
+    """Defining base class for inheritence."""
+
+    def help(self):
+        """Return all available commands."""
+        return """
+        cmnds
+        ----
+        !help
+        !version
+        !rand
+        !flip
+        !chuck
+        !gif
+        !haiku
+        !names
+        !hn
+        !twitch
+        !bored
+        !trivia
+        !gnews
+        """
