+++ signalbot/functions/me.py
@@ -1,0 +1,12 @@
+"""Me module."""
+
+import emoji
+
+
+class Me:
+    """Defining base class for inheritence."""
+
+    def me(self):
+        """Return random Emoij."""
+        thumb = emoji.emojize(':eggplant:')
+        return thumb
