+++ signalbot/functions/test.py
@@ -1,0 +1,9 @@
+"""Test module."""
+
+
+class Test:
+    """Defining base class for inheritence."""
+
+    def test(self):
+        """Simple test function."""
+        return "@#*&ES&@#YF.. nooo you got me!"
