+++ signalbot/functions/winamp.py
@@ -1,0 +1,12 @@
+"""Winamp module."""
+
+import emoji
+
+
+class Winamp:
+    """Defining base class for inheritence."""
+
+    def winamp(self):
+        """Return random Emoij."""
+        thumb = emoji.emojize(':llama:')
+        return "It really whips the " + thumb + " ass."
