+++ signalbot/metadata.py
@@ -1,0 +1,3 @@
+author = "Patrick Blaas"
+authoremail = "<patrick@kite4fun.nl>"
+version = "0.1.4"
